==> notebooks/2_API_scraping_open_libarary.ipynb <==
import pandas as pd
import numpy as np
import requests
import json
import time
from datetime import datetime, timedelta

## Helper functions

import requests
import pandas as pd

def get_books(urls):
    books_ol = []

    for url in urls:
        print(f"Fetching from: {url}")
        try:
            response = requests.get(url)
            response.raise_for_status()
            response_dict = response.json()
        except Exception as e:
            print("Error fetching this URL:", e)
            continue

        for b in response_dict.get("works", []):
            try:
                title = b.get("title", "Unknown Title")
                author = b.get("author_name", ["Unknown Author"])[0]
                year_published = b.get("first_publish_year", None)

                # Fetch ratings
                url_rating = f"https://openlibrary.org/search.json?title={title.replace(' ', '+')}&fields=*"
                response_rating = requests.get(url_rating)
                response_rating_dict = response_rating.json()
                avg_rating = response_rating_dict["docs"][0].get("ratings_average", None)
                rating_count = response_rating_dict["docs"][0].get("ratings_count", None)

                # Fetch subjects / genres
                url_book = f"https://openlibrary.org{b['key']}.json"
                response_book = requests.get(url_book)
                response_book_dict = response_book.json()
                genres = response_book_dict.get("subjects", [])

                for g in genres:
                    books_ol.append({
                        "title": title,
                        "author": author,
                        "avg_rating": avg_rating,
                        "rating_count": rating_count,
                        "year_published": year_published,
                        "genre": g
                    })

            except Exception as e:
                print("there was an error so we continue with the next book:", e)
                continue

    df = pd.DataFrame(books_ol)
    return df


# def get_books(urls):

#     books_ol = []
    
#     for url in urls:
#         print('HI!')
#         for b in response_dict["works"]:
#             try:
#                 author = b['author_name'][0]
#                 title = b['title']
#                 print(title)
#                 year_published = b['first_publish_year']
#                 url_rating = r'https://openlibrary.org/search.json?title='+ title.replace(' ', '+') + '&fields=*'
#                 response_rating = requests.get(url_rating)
#                 response_rating_dict = response_rating.json()
#                 avg_rating = response_rating_dict['docs'][0]['ratings_average']
#                 rating_count = response_rating_dict['docs'][0]['ratings_count']
#                 url_book = r'https://openlibrary.org' + b['key'] + '.json'
#                 response_book = requests.get(url_book)
#                 response_book_dict = response_book.json()
#                 genres = response_book_dict['subjects']
#                 for g in genres:
#                     books_ol.append({'title':title,
#                                      'author':author,
#                                      'avg_rating':avg_rating,
#                                      'rating_count':rating_count,
#                                      'year_published':year_published,
#                                      'genre':g
#                                     })
#             except Exception as e:
#                 print('there was an error so we continue with the next book:', e)
#                 continue
#     df = pd.DataFrame(books_ol)
#     return df     


## Extract data

url_list = ["https://openlibrary.org/trending/yearly.json?offset=0&page=1",
        "https://openlibrary.org/trending/yearly.json?offset=0&page=2",
        "https://openlibrary.org/trending/yearly.json?offset=0&page=3",
        "https://openlibrary.org/trending/yearly.json?offset=0&page=4",
        "https://openlibrary.org/trending/yearly.json?offset=0&page=6",
        "https://openlibrary.org/trending/yearly.json?offset=0&page=7",
        "https://openlibrary.org/trending/yearly.json?offset=0&page=8",
        "https://openlibrary.org/trending/yearly.json?offset=0&page=9",
        "https://openlibrary.org/trending/yearly.json?offset=0&page=10"
       ]

df_books_ol = get_books(url_list) 


print(type(df_books_ol))
print(df_books_ol.head(20))


df_books_ol["title"].nunique()

df_books_ol_raw = df_books_ol.to_csv("../data/raw/books_ol_raw.csv",index=False)

# url = r'https://openlibrary.org/trending/yearly.json?offset=0&page=1'
# response = requests.get(url)
# response_dict = response.json()
# response_dict.keys()

# type(response_dict["works"])

# len(response_dict["works"])

# response_dict["works"][0]

# author = response_dict['works'][0]['author_name'][0]
# title = response_dict['works'][0]['title']
# year_published = response_dict['works'][0]['first_publish_year']

# # we create the HTML address based on info available on the website
# url_book = r'https://openlibrary.org' + response_dict['works'][0]['key'] + '.json'
# print(url_book)
# response_book = requests.get(url_book)
# response_book_dict = response_book.json()

# response_book_dict.keys()

# response_book_dict['subjects']



# url_rating = r'https://openlibrary.org/search.json?title='+ response_dict['works'][0]['title'].replace(' ', '+') + '&fields=*'
# url_rating



# response_rating = requests.get(url_rating)
# response_rating_dict = response_rating.json()

# response_rating_dict.keys()

# avg_rating = response_rating_dict['docs'][0]['ratings_average']
# rating_count = response_rating_dict['docs'][0]['ratings_count']
